clean src/lib/types.ts
