+++ src/lib/index.ts
@@ -17,7 +17,6 @@ export {
   type CCIPExecution,
   type CCIPMessage,
   type CCIPRequest,
-  type CCIPSupportedToken,
   type CCIPVersion,
   type CommitReport,
   type EVMExtraArgsV1,
@@ -25,7 +24,6 @@ export {
   type ExecutionReceipt,
   type Lane,
   type NetworkInfo,
-  type PoolSupportCheck,
   CCIPContractTypeCommitStore,
   CCIPContractTypeOffRamp,
   CCIPContractTypeOnRamp,
