+++ src/lib/types.ts
@@ -178,23 +178,3 @@ export function parseSourceTokenData(data: string): SourceTokenData {
   const decoded = defaultAbiCoder.decode([SourceTokenData], data)
   return (decoded[0] as Result).toObject() as SourceTokenData
 }
-
-// Basic token metadata (standard ERC20 metadata)
-export interface TokenMetadata {
-  name: string
-  symbol: string
-  decimals: bigint
-}
-
-// Extended token info with pool details for CCIP
-export interface CCIPSupportedToken extends TokenMetadata {
-  address: string
-  pool: string
-}
-
-export interface PoolSupportCheck {
-  token: string
-  pool: string
-  isSupported: boolean
-  error: Error | null
-}
